requirements.txt
# Core
openai>=1.0.0
python-dotenv

# Web UI (Streamlit)
streamlit

